requirements.txt
streamlit
pandas
plotly
joblib
numpy
groq
python-dotenv
scikit-learn